Rename overlap.py as soss_engine.py
