+++ SOSS/extract/Empirical_Trace/construct_trace.py
@@ -17,7 +17,7 @@ import webbpsf
 from SOSS.trace import tracepol as tp
 from SOSS.extract.empirical_trace import centroid as ctd
 from SOSS.extract.empirical_trace import plotting as plotting
-from SOSS.extract.overlap import TrpzOverlap, TrpzBox
+from SOSS.extract.soss_engine import TrpzOverlap, TrpzBox
 from SOSS.extract.throughput import ThroughputSOSS
 from SOSS.extract.convolution import WebbKer
